RSI with SMA
This is Ram Uncle's formula. I'm trying to get his green signal to use it. Please note that this works only when the window is more than 65 days, so you will have to go back to November for this. You have to call Ticker again to get new scloses and sdates.